+++ tools/rpkg/CMakeLists.txt
@@ -20,6 +20,7 @@ find_library(
 
 include_directories(${R_INCLUDE_DIR})
 include_directories(../../)
+include_directories(src/include)
 
 # this is used for clang-tidy checks and IDE support
 add_library(
